backend/requirements.txt
fastapi
uvicorn
pdfminer.six
pandas
openpyxl
python-multipart
pydantic
python-docx

